requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.2.4
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.11.6
djangorestframework==3.7.3
dj-database-url==0.4.2
dj-static==0.0.6
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pip==9.0.1
psycopg2==2.7.3.2
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
setuptools==38.2.4
static3==0.7.0
uritemplate==3.0.0
urllib3==1.22
wheel==0.30.0
